// test_data/parameters.enum.ts
export enum DurationEnum {
    "1 year" = "1 Year",
    "2 years" = "2 Years",
    "3 years" = "3 Years",
}

export enum PurchaseType {
    Subscription = "Subscription",
    Perpetual = "Perpetual",
    Single = "Single",
    Team = "Team",
    Site = "Site",
}

export enum Edition {
    Standard = "Standard",
    Professional = "Professional",
    Enterprise = "Enterprise",
    Developer = "Developer",
    "SSIS Integration Database Bundle" = "SSIS Integration Database Bundle",
}